HUE-7388 [core] Include schedule job counts into the global job count widget
The running job conter is now split into individual counters: one for YARN jobs,
one for Scheduler jobs... That way each one can be updated individually and new ones
(e.g. Impala queries...) can be easily added.

The status of the integrated scheduler is also pulled on the right assist Scheduler page.
Bit more polishing there to do on the UI but the goal is to add the logic/skeleton first
and iterate.

And now we have individual calls:
        huePubSub.subscribe('check.job.browser', checkYarnBrowserStatus);
        huePubSub.subscribe('check.schedules.browser', checkScheduleBrowserStatus);
        ... queries next.

And the global counter ones:
  $.when.apply($, [lastYarnBrowserRequest, lastScheduleBrowserRequest])